rename native/external to public
