+++ cmd/tools/cassandra/main.go
@@ -23,7 +23,7 @@ package main
 import (
 	"os"
 
-	_ "github.com/uber/cadence/common/persistence/nosql/nosqlplugin/cassandra/gocql/native" // needed to load the default gocql client
+	_ "github.com/uber/cadence/common/persistence/nosql/nosqlplugin/cassandra/gocql/public" // needed to load the default gocql client
 	"github.com/uber/cadence/tools/cassandra"
 )
